Added RouteId Property to GameInfo
This was added because the website url cannot show spaces in game titles. Instead of keeping the space and making it hard for the user to read, I replaced all spaces in the title with a dash for use as an id in the url. CanonicalGames, and anchors have been updated accordingly.