Updated Home/Index & Added Missing Title to Games/Detail
This isn't related to the current issue; but I'm a team of 1 person, and I prefer to fix small problems like this when I see them; as opposed to making a separate issue/branch just to change two lines. If I had other people on this team, I would create a new branch/issue.