Initial Implementation of Server-Side Connect Four
Some of this code is a real nightmare. I suspect it will not function without some corrections. The client-side stuff needs to be implemented to ensure that the system will work in general; but also to test for bugs (especially in the spots where I used Reflection with SignalR contexts).