Slightly Refactor Shared Match Forfeit Functionality for Games
The shared handling wasn't declaring that the game ended to the clients.

I changed some calls to reset the game's client to use the ResetCallerGame method. The method is virtual, so it should be used instead of a direct RPC to the client's ResetGame function; otherwise, it could result in bugs if a future game hub overrides ResetCallerGame.

I added a method to end the match, and to set the player turn externally. For consistency with the existing external helper methods that have direct access to the hub, I made them non-static. However, these should be investigated; as I think SignalR may allow the client to call them. A redesign for external helper methods may be necessary. They only exist to keep SignalR RPC calls outside of game code.