Identify and Fix Bug Related to Statistics Tracking in Connect Four
I noticed that statistics weren't being tracked for Checkers in testing; and, forgot that it just hadn't been implemented, yet. On a whim, I checked Connect Four, and it also had a similar issue. I thought it could be related to shared functionality; but, in the end, it was a misplaced call in Connect Four's handling.

This fix is unrelated to the current branch; but, as I am a team of one, I will shortcut the usual issue -> branch -> fix -> PR -> review -> merge process and just fix it here.